LayoutTests/fast/images/image-copy-memory-usage-expected.txt

This is an automated test related to Bug 152374 - Reduce resource usage when copying image to clipboard.

This tests how much memory is used while copying an animation into the clipboard. The animation consists of 50 500x500 frames and, assuming 32 bits per pixel, will use 50 MB of memory when uncompressed. To pass the test, after copying the animation to the clipboard, the memory usage of the browser must not have increased by more than 200 MBs.

In order to make this test more accurate it should be performed in a freshly started browser session that has nothing else open and which has an empty cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS changeInCommittedVMBytes < 200000000 is true
PASS successfullyParsed is true

TEST COMPLETE
